resource-proxy: first version of resource-proxy.
Resource-proxy allows one Murphy instance to be a "slave" to another Murphy: the
resource requests are proxied to the master. The connection creation and library
initialization is done using resource-proxy plugin. Since the Murphys can reside
even in different computers, the zone is overwritten to match what is given as
argument to the plugin.

This version has several things still under construction. An example would be
the way how the libraries are linked at the moment.