resource-proxy: fix a memory leak.
A bigger memory leak still remains: we need to figure out when it's safe to
remove association between a sequence number to the actual request that
originated it. The problem is that the protocol calls events with the same
sequence number also after the procedure call is over (the response has been
received), so it's difficult to say when there will be no more events with the
same sequence number.