common: support cookie-only hash-table operations.
Add support for looking up and deleting hash table entries based on
cookies only. You can pass to the affected functions MRP_HASH_KEY_NONE
as the key to trigger this mode of operation. For cookie-only lookups
you can also use the newly introduced mrp_hashtbl_fetch macro.

As a side-effect, mrp_hashtbl_replace has been changed to be more in
line with mrp_hashtbl_add. You need to pass in a pointer to the cookie
or NULL in case you're not using cookies. If you do use them, set the
pointed to cookie to MRP_HASH_COOKIE_NONE if you don't have one, in
which case upon return it will be updated to contain the cookie of the
new/replaced entry.

Also tried to clean up the code, to make it a bit more self-explanatory,
get rid of scattered repetitive blocks, get rid of unnecessary tests,
etc. It's still not quite where I'd like it to be, so there should be
more of the same to come.